Don't delete SOA records

+++ sync_dns.py
@@ -184,7 +184,7 @@ def sync_zone(domain_records_url, domain):
     # Delete any records that exist with DigitalOcean that have been removed
     print "\nRemoving deleted records"
     for record in existing_records:
-        if record['id'] not in synced_record_ids:
+        if record['id'] not in synced_record_ids and record['type'] != 'SOA':
             response = requests.delete("{0}/{1}".format(domain_records_url, record["id"]), headers=headers)
             if response.status_code == 204:
                 print "--> Deleted record", record["name"], "IN", record["type"], record["data"]
